// Angular-Playground/platform/projects/rxjs/src/lib/models/film.ts
import { StarWarsType } from "./responseWrapper";

export interface Film {
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  characters: string[];
  planets: string[];
  starships: string[];
  vehicles: string[];
  species: string[];
  created: string;
  edited: string;
  url: string;
  type: StarWarsType;
}